# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.6.0
  description: |
    Seshat API — Plataforma de gestão educacional com registro fotográfico, equipes, turmas e ocorrências.

    ## Banco de dados

    O PostgreSQL 16 roda **dentro do container de desenvolvimento** e é gerenciado
    automaticamente pelo `entrypoint.sh`:

    1. **Inicialização do cluster** — `pg_createcluster` é executado somente se o
       volume estiver vazio (`PG_VERSION` ausente em `PGDATA`).
    2. **Persistência** — dados armazenados no volume Docker `carometro-pg-dev`
       montado em `/var/lib/postgresql/16/main`.
    3. **Schema** — `drizzle-kit push` roda automaticamente **somente quando** o
       hash dos arquivos `lib/db/src/**/*.ts` muda. O hash é persistido em
       `$PGDATA/.carometro_schema_hash`.
    4. **Seed do administrador** — executado **apenas na primeira inicialização**
       via marcador `$PGDATA/.carometro_initialized`. O script `seed-admin.ts` é
       idempotente (não duplica o usuário se já existir).

    ### Comandos úteis

    ```
    docker compose run --rm dev db:push       # forçar reaplicação do schema
    docker compose run --rm dev db:seed       # verificar/recriar administrador
    docker compose run --rm dev db:hash-reset # remover marcadores e forçar re-setup
    docker compose down -v && docker compose up --build  # reinício limpo
    ```
servers:
  - url: /api
...
paths:
...
  /import/disciplinas:
    post:
      operationId: importDisciplinas
      tags: [import]
      summary: Import disciplinas and their ofertas from spreadsheet rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportDisciplinasBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /usuarios/{id}/disciplinas:
    put:
      operationId: setUsuarioDisciplinas
      tags: [usuarios]
      summary: Set disciplinas for a usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDisciplinasInput"
      responses:
        "200":
          description: Updated usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
...
  /disciplinas:
    get:
      operationId: listDisciplinas
      tags: [disciplinas]
      summary: List all disciplinas
      responses:
        "200":
          description: List of disciplinas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Disciplina"
    post:
      operationId: createDisciplina
      tags: [disciplinas]
      summary: Create a disciplina
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplinaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disciplina"
  /disciplinas/{id}:
    put:
      operationId: updateDisciplina
      tags: [disciplinas]
      summary: Update a disciplina
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplinaInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disciplina"
    delete:
      operationId: deleteDisciplina
      tags: [disciplinas]
      summary: Delete a disciplina
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
  /disciplinas/{id}/ofertas:
    put:
      operationId: setDisciplinaOfertas
      tags: [disciplinas]
      summary: Set cursos and turnos (ofertas) for a disciplina
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDisciplinaOfertasInput"
      responses:
        "200":
          description: Updated disciplina with ofertas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disciplina"
components:
  schemas:
...
    UsuarioDisciplina:
      type: object
      required: [ofertaId, disciplinaId, disciplinaNome, cursoId, cursoNome, turnoId, turnoNome]
      properties:
        ofertaId:
          type: string
          format: uuid
        disciplinaId:
          type: string
          format: uuid
        disciplinaNome:
          type: string
        cursoId:
          type: string
          format: uuid
        cursoNome:
          type: string
        turnoId:
          type: string
          format: uuid
        turnoNome:
          type: string
...
    ImportRow:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true
...
    ImportDisciplinasBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
...
    ImportResult:
      type: object
      required: [imported, errors]
      properties:
        imported:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    Usuario:
      type: object
      required: [id, email, codigoAcesso, primeiroAcesso, fotoUrl, roles, disciplinas, criadoEm]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
          nullable: true
        email:
          type: string
        dataNascimento:
          type: string
          format: date
          nullable: true
          description: Data de nascimento (YYYY-MM-DD). Obrigatória quando o perfil "estudante" for atribuído a menores de 18 anos.
        codigoAcesso:
          type: string
        primeiroAcesso:
          type: boolean
        fotoUrl:
          type: string
          nullable: true
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleSummary"
          description: >
            Um usuário pode ter múltiplos perfis simultaneamente.
            Exceção: usuários com o perfil "estudante" e menores de 18 anos
            só podem ter esse único perfil — nenhum outro pode ser combinado.
        disciplinas:
          type: array
          items:
            $ref: "#/components/schemas/UsuarioDisciplina"
        criadoEm:
          type: string
    RoleSummary:
      type: object
      required: [id, nome]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
...
    Disciplina:
      type: object
      required: [id, nome, criadoEm, ofertas]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        criadoEm:
          type: string
        ofertas:
          type: array
          items:
            $ref: "#/components/schemas/DisciplinaOferta"
    DisciplinaOferta:
      type: object
      required: [id, disciplinaId, cursoId, cursoNome, turnoId, turnoNome]
      properties:
        id:
          type: string
          format: uuid
        disciplinaId:
          type: string
          format: uuid
        cursoId:
          type: string
          format: uuid
        cursoNome:
          type: string
        turnoId:
          type: string
          format: uuid
        turnoNome:
          type: string
    DisciplinaInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          minLength: 1
    SetDisciplinaOfertasInput:
      type: object
      required: [ofertas]
      properties:
        ofertas:
          type: array
          items:
            type: object
            required: [cursoId, turnoId]
            properties:
              cursoId:
                type: string
                format: uuid
              turnoId:
                type: string
                format: uuid
    SetDisciplinasInput:
      type: object
      required: [disciplinaOfertaIds]
      properties:
        disciplinaOfertaIds:
          type: array
          items:
            type: string
            format: uuid
...